// src/app/models/addclaim.ts
export interface addclaim {
    claimId: string,
    customerName: string;
    customerEmail: string; 
    vehicleMake: string;
    vehicleModel: string;
    modelYear: string;
    MMCode: string;
    damageDescription: string;
    customerSurbub: string;
    customerCity: string;
    customerProvince: string;
    claimsAgent: string;
    claimsAgentId: string; 
}